fix(metadata): add metadata to certificate lookup and provider-apply pages
Both pages were "use client" components with no metadata export, causing
the metadata coverage test to fail. Extracted client logic to *Client.tsx
files so the page.tsx can be a server component with metadata exports.
